voegt comments toe
