+++ imports/ui/body.html
@@ -3,7 +3,7 @@
     <header>
       <h1>Truck Order List</h1>
     </header>
-
-{{> reactiveTable class="table table-bordered table-hover" settings=settings}}
-</div>
+    <!-- Template for the table  -->
+    {{> reactiveTable class="table table-bordered table-hover" settings=settings}}
+  </div>
 </body>
