+++ imports/ui/body.js
@@ -3,14 +3,4 @@ import { TruckOrders } from '../api/truckOrders.js';
 
 import './body.html';
 
-Template.body.helpers({
-  settings: function () {
-    var collection = TruckOrders.find({ arrival_time: { $exists: true, $ne: null}});
-    return {
-      collection: collection,
-      rowsPerPage: 10,
-      showFilter: true,
-      fields: ['t_number', 'arrival_time', 'number']
-    };
-  }
-});
+import './table.js';
